Adds an option to do a dry-run.

+++ main.py
@@ -1,12 +1,13 @@
+from arguments import parser
 from utils import *
 
 
-def main():
-    calendar = get_calendar_service()
+def main(dry: bool = False):
+    calendar = get_calendar_service() if not dry else None
 
-    remote_appointments: dict = load_cached_appointments_from_disk()
+    remote_appointments: dict = get_stored_appointments()
     local_appointments: dict = {}
-    appointments_in_workbook: list = load_appointments_from_workbook()
+    appointments_in_workbook: list = get_appointments_from_workbook()
 
     for appointment in appointments_in_workbook:
         checksum = appointment.checksum()
@@ -15,17 +16,25 @@ def main():
             local_appointments[checksum] = remote_appointments[checksum]
             remote_appointments.pop(appointment.checksum())
         else:
-            event_id = create_appointment(calendar, appointment)
-            if event_id is not None:
-                local_appointments[appointment.checksum()] = event_id
+            if not dry:
+                event_id = create_appointment(calendar, appointment)
+                if event_id is not None:
+                    local_appointments[appointment.checksum()] = event_id
+            else:
+                print("Would create an new event " + appointment.title)
 
     for uid in remote_appointments.values():
-        delete_appointment(calendar, uid)
+        if not dry:
+            delete_appointment(calendar, uid)
+        else:
+            print("Would delete an event " + uid)
 
-    with open(STORAGE_FILE, 'w') as f:
-        for checksum, uid in local_appointments.items():
-            f.write(f'{checksum} {uid}\n')
+    if not dry:
+        with open(STORAGE_FILE, 'w') as f:
+            for checksum, uid in local_appointments.items():
+                f.write(f'{checksum} {uid}\n')
 
 
 if __name__ == '__main__':
-    main()
+    options = parser.parse_args()
+    main(options.dry)
